+++ bfabric/docs/changelog.md
@@ -10,10 +10,24 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+### Breaking
+
+- `base_url` is now a mandatory parameter in the configuration.
+
+### Added
+
+- It is now possible to configure `~/.bfabricpy.yml` without a default environment. In that case it will always be
+    necessary to specify the requested config environment to be used.
+- `bfabric.entities.Executable` has `parameters` relationship now
+
+### Removed
+
+- Some old submitter related functionality is deleted.
+
 ### Changed
 
 - Columns of tables named after B-Fabric entities, containing only integers, will be set as the specified type
-    when saving to B-Fabric.
+    when saving to B-Fabric (in `experimental.upload_dataset`).
 
 ## \[1.13.27\] - 2025-05-21